File split the calendar.js file into addTask.js, calendar.js, listView.js, styles.js, and taskManagment.js. Fixed an issue with the edit modal not edditing the correct task. Fixed a little styling issues/ cleaned up some of the looks.
