+++ app/_layout.tsx
@@ -1,3 +1,5 @@
+// app/_layout.tsx
+
 import { Stack } from 'expo-router';
 
 export default function Layout() {
