+++ components/Collapsible.tsx
@@ -1,3 +1,5 @@
+// components/Collapsible.tsx
+
 import { PropsWithChildren, useState } from 'react';
 import { StyleSheet, TouchableOpacity } from 'react-native';
