+++ components/HapticTab.tsx
@@ -1,3 +1,5 @@
+//components/HapticTab.tsx
+
 import { BottomTabBarButtonProps } from '@react-navigation/bottom-tabs';
 import { PlatformPressable } from '@react-navigation/elements';
 import * as Haptics from 'expo-haptics';
