+++ components/ParallaxScrollView.tsx
@@ -1,3 +1,5 @@
+//components/ParallaxScrollView.tsx
+
 import type { PropsWithChildren, ReactElement } from 'react';
 import { StyleSheet } from 'react-native';
 import Animated, {
