+++ components/ThemedText.tsx
@@ -1,6 +1,6 @@
+// components/ThemedText.tsx
 import { Text, type TextProps, StyleSheet } from 'react-native';
-
-import { useThemeColor } from '@/hooks/useThemeColor';
+import { useThemeColor } from '../hooks/useThemeColor';
 
 export type ThemedTextProps = TextProps & {
   lightColor?: string;
@@ -9,27 +9,26 @@ export type ThemedTextProps = TextProps & {
 };
 
 export function ThemedText({
-  style,
-  lightColor,
-  darkColor,
-  type = 'default',
-  ...rest
-}: ThemedTextProps) {
+                             style,
+                             lightColor,
+                             darkColor,
+                             type = 'default',
+                             ...rest
+                           }: ThemedTextProps) {
   const color = useThemeColor({ light: lightColor, dark: darkColor }, 'text');
-
   return (
-    <Text
-      style={[
-        { color },
-        type === 'default' ? styles.default : undefined,
-        type === 'title' ? styles.title : undefined,
-        type === 'defaultSemiBold' ? styles.defaultSemiBold : undefined,
-        type === 'subtitle' ? styles.subtitle : undefined,
-        type === 'link' ? styles.link : undefined,
-        style,
-      ]}
-      {...rest}
-    />
+      <Text
+          style={[
+            { color },
+            type === 'default' ? styles.default : undefined,
+            type === 'title' ? styles.title : undefined,
+            type === 'defaultSemiBold' ? styles.defaultSemiBold : undefined,
+            type === 'subtitle' ? styles.subtitle : undefined,
+            type === 'link' ? styles.link : undefined,
+            style,
+          ]}
+          {...rest}
+      />
   );
 }
